+++ srcs/parsing/get_arguments.c
@@ -6,7 +6,7 @@
 /*   By: jbarbay <jbarbay@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/11 18:09:13 by jbarbay           #+#    #+#             */
-/*   Updated: 2024/04/25 12:47:07 by jbarbay          ###   ########.fr       */
+/*   Updated: 2024/04/25 14:42:34 by jbarbay          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -98,9 +98,9 @@ void	get_textures_and_colors(t_game_data *data)
 		check_identifier(data, args, line);
 		line = prepare_next_iteration(line, args, data->fd, 0);
 	}
-	check_xpm(data->north_path, data, line);
-	check_xpm(data->south_path, data, line);
-	check_xpm(data->west_path, data, line);
-	check_xpm(data->east_path, data, line);
+	check_extension(data->north_path, line, data, ".xpm");
+	check_extension(data->south_path, line, data, ".xpm");
+	check_extension(data->west_path, line, data, ".xpm");
+	check_extension(data->east_path, line, data, ".xpm");
 	free(line);
 }
