+++ srcs/parsing/validation_helpers.c
@@ -6,7 +6,7 @@
 /*   By: jbarbay <jbarbay@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/08 14:32:43 by jbarbay           #+#    #+#             */
-/*   Updated: 2024/04/25 13:07:48 by jbarbay          ###   ########.fr       */
+/*   Updated: 2024/04/25 14:49:09 by jbarbay          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -61,26 +61,17 @@ void	error_parsing(char *message, char **array, char *line, t_game_data *data)
 	exit(1);
 }
 
-void	check_cub_file(char *filename)
+void	check_extension(char *file, char *line, t_game_data *data, char *ext)
 {
 	int	length;
 	int	result;
 
-	length = ft_strlen(filename);
-	filename += (length - 4);
-	result = ft_strncmp(filename, ".cub", 4);
-	if (result != 0)
-		error_parsing("Please provide all the arguments in a .ber file", NULL, NULL, NULL);
-}
-
-void	check_xpm(char *filename, t_game_data *data, char *line)
-{
-	int	length;
-	int	result;
-
-	length = ft_strlen(filename);
-	filename += (length - 4);
-	result = ft_strncmp(filename, ".xpm", 4);
-	if (result != 0)
+	length = ft_strlen(file);
+	file += (length - 4);
+	result = ft_strncmp(file, ext, 4);
+	if (result != 0 && ext[1] == 'x')
 		error_parsing("Please provide textures in XPM format", NULL, line, data);
+	else if (result != 0)
+		error_parsing("Please provide the map in a .cub file", NULL, line, data);
 }
+
