+++ test.cub
@@ -12,7 +12,7 @@ C  124,185,232
 11000000111101
 10000000000011
 1000000N000001
-10000000000000001
+100000000000001
 11000000000001
 10000000000001
 10000000000001
