+++ textures/eagle.xpm
